requirements.txt
# ── Core Framework ──
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic[email]==2.11.2
pydantic-settings==2.9.1

# ── Database ──
sqlalchemy==2.0.40
aiosqlite==0.21.0

# ── Authentication ──
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.3.0

# ── Document Parsing ──
PyMuPDF==1.25.5
python-pptx==1.0.2

# ── AI / ML ──
sentence-transformers==4.1.0
faiss-cpu==1.11.0
httpx==0.28.1

# ── Audio ──
gTTS==2.5.1
openai-whisper

# ── Utilities ──
python-multipart==0.0.20
python-dotenv==1.1.0